refactor: move search options
