+++ src/webview/SearchSidebar/SearchWidgetContainer/index.tsx
@@ -1,7 +1,5 @@
-import { SearchInput } from './comps/SearchInput'
-import { Center, HStack, VStack } from '@chakra-ui/react'
-import { useBoolean } from 'react-use'
-import { VscChevronRight, VscChevronDown } from 'react-icons/vsc'
+import SearchOptions from './comps/SearchOptions'
+import SearchWidget from './comps/SearchWidget'
 
 interface Props {
   inputValue: string
@@ -14,38 +12,15 @@ const SearchWidgetContainer = ({
   setInputValue,
   refreshResult
 }: Props) => {
-  const [isExpanded, toggleIsExpanded] = useBoolean(false)
   return (
-    <HStack position="relative">
-      <Center
-        w={16}
-        h="100%"
-        cursor="pointer"
-        position="absolute"
-        top="0"
-        left="2px"
-        onClick={toggleIsExpanded}
-        _hover={{
-          background: 'var(--vscode-toolbar-hoverBackground)'
-        }}
-      >
-        {isExpanded ? <VscChevronDown /> : <VscChevronRight />}
-      </Center>
-      <VStack gap={6} flex={1} ml="18px">
-        <SearchInput
-          value={inputValue}
-          onChange={setInputValue}
-          onKeyEnterUp={refreshResult}
-        />
-        {isExpanded ? (
-          <SearchInput
-            value={inputValue}
-            onChange={setInputValue}
-            onKeyEnterUp={refreshResult}
-          />
-        ) : null}
-      </VStack>
-    </HStack>
+    <div>
+      <SearchWidget
+        inputValue={inputValue}
+        setInputValue={setInputValue}
+        refreshResult={refreshResult}
+      />
+      <SearchOptions />
+    </div>
   )
 }
