+++ src/webview/SearchSidebar/index.tsx
@@ -6,7 +6,6 @@ import { useSearchResult } from '../hooks/useSearch'
 import LoadingBar from '../LoadingBar'
 import SearchProviderMessage from './SearchProviderMessage'
 import { useDeferredValue } from 'react'
-import SearchOptions from './SearchOptions'
 
 export const SearchSidebar = () => {
   const [inputValue = '', setInputValue] = useLocalStorage(
@@ -35,7 +34,6 @@ export const SearchSidebar = () => {
         refreshResult={refreshSearchResult}
         setInputValue={setInputValue}
       />
-      <SearchOptions />
       <SearchProviderMessage
         pattern={queryInFlight}
         error={searchError}
